break ScrollerController into its own module
