+++ .prettierignore
@@ -280,6 +280,10 @@ packages/labs/virtualizer/LitVirtualizer.js
 packages/labs/virtualizer/LitVirtualizer.d.ts
 packages/labs/virtualizer/LitVirtualizer.d.ts.map
 packages/labs/virtualizer/LitVirtualizer.js.map
+packages/labs/virtualizer/ScrollerController.js
+packages/labs/virtualizer/ScrollerController.d.ts
+packages/labs/virtualizer/ScrollerController.d.ts.map
+packages/labs/virtualizer/ScrollerController.js.map
 packages/labs/virtualizer/test/**/*.d.ts
 packages/labs/virtualizer/test/**/*.d.ts.map
 packages/labs/virtualizer/test/**/*.js
